main.py파일 위치 변경
변경 내용: wumpus-world/wumpus/main.py에서,
wumpus-world/main.py으로 위치 변경.

이유: 디렉터리 구조 유지, ModuleNotFoundError해결 위함
